requirements.txt
aiohttp>=3.9,<4
jinja2>=3.1,<4
lxml>=5.0,<6
pyyaml>=6.0,<7
